Add Either#to_either
This makes it easy to consume any type of monad object that is convertible to Either, including Either objects themselves.

More practically speaking, when we’re working with objects from dry-monads, we can just call `#to_either` on it without checking for a `respond_to?`.

This helps make our matching in dry-result_matcher more straightforward.