Rename Try's constructors
This commit renames Try::Success to Try::Value and Try::Failure to Try::Error.
The reason behind this change is quite simple, we renamed Either's constructors
to Success and Failure and for keeping easier to understand it's better not to
have same names for monadic values of different types. I didn't any backward
compatibility except for the `success?` and `failure?` aliases because the Try
monad in fact doesn't suppose constructing values directly.