Introduce Result#either method
`Result#either` takes two callable arguments. If the value is a success
passes it as argument to the first function and returns the result. If it
is a failure does the same with the second function.

```ruby
Dry::Monads.Success(1).either(-> x { x + 1 }, -> x { x + 2 }) # => 2
Dry::Monads.Failure(1).either(-> x { x + 1 }, -> x { x + 2 }) # => 3
```