Add Maybe#maybe
At the moment, Maybe#fmap violates the functor laws because it coerces nil values to None. This ruins predictability of composition. Changing current behavior would be a breaking change so I won't do this. This commit instead adds `Maybe#maybe`, an alias for `fmap` that does coercing too. In future, calling `fmap` with a block returning `nil` will issue a warning, in 2.0 this warning will be turned into an error.