Fix keyword warnings
This changes rules of destructuring values since starting Ruby 2.7 keyword arguments are allowed to be arbitrary objects. I simplified the check for now to make it semi-compatible but I will remove it in the next major version. Programming with do notation is preferrable and it doesn't require all these weird stuff I invented back in the dry-transaction days.