Add note about Do::All caveat
When Do::All is included, it specifically excludes methods inherited from ancestors such as included modules or parent classes. This is likely very intentional to avoid wrapping methods right up the hierarchy. I discovered this limitation when putting some shared logic in a module that two classes using Do::All could include, to discover that it doesn't work. A workaround is to instead have the module class_eval the instance methods onto the class.